requirements.txt
fastapi
aiologger[aiofiles]
pip-chill
uvicorn
gunicorn
langchain-community
langgraph
tavily-python
langchain-openai
langchain
bs4
chromadb
tiktoken
openai
langchain_text_splitters
markdownify